requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
async-timeout==5.0.0
attrs==24.2.0
black==24.10.0
cffi==1.17.1
click==8.1.7
colorama==0.4.6
discord.py==2.4.0
flake8==7.1.1
frozenlist==1.5.0
idna==3.10
isort==5.13.2
mccabe==0.7.0
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
propcache==0.2.0
py-cord==2.6.1
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
yarl==1.17.1
